Disable a test on clang 3.6 due to clang bug
QStringList foo = { "foo" }; in this case, even though the ast dump
looks the same in both clang 3.6 and clang 3.7 the CXXConstructorExpr
which calls the QString ctor isn't being visited with clang 3.6, which
is very weird because it's in the ast tree.

Blacklist the test on older clang for now. Planing on writing a proper
test system which is more configurable, so we could blacklist only a few lines
instead of the whole test.