tests: Allow to specify which .cpp files to build through json
So, instead of building *cpp we can build individual files. This opens
up lots of possibilities, like conditionally building some files depending
on clang version.

Currently we have some tests blacklisted on clang 3.6 just because of 1% of the test,
now it will be possible to move that 1% to a separate file and still have coverage
on the other 99%