Add a way of accessing the last statement inside VisitDecl
This will help solve the problem of determining if we're inside a try-catch block.

try {
} catch (Foo foo) {
}

It's usually easy to go up the AST and check for a CXXTryStmt or CXXCatchStmt,
but if you're visiting 'foo' VarDecl you won't have any parent statements
because the tree is "rooted" on the decl.

As a workaround, you can use this (inside VisitDecl()):

    bool isInCatch = isa<CXXCatchStmt>(m_lastStmt);

The statements that compose the catch parameters don't have a parent,
you can't navigate up the AST