reserve-candidates: Don't warn for QVarLengthArray
Unlike the other containers QVLA doesn't need reserve in the majority
of cases since it uses the stack, resulting in false-positives.

There's still the edge case of it using the heap, but that would
take a considerable amount of effort to catch, and besides, users
using QVLA usually know what they are doing.

BUG: 367484