connect-non-signal: Don't warn if the method decl comes from a macro
This is very hard to support, since both the method and the signal macro
will have the same source location. Maybe it can be done by falling back to
comparing the spelling location whenever the locs are the same, but would mean
a very large rewrite

Anyway, someone using a macro that spits "Q_SIGNALS: void myMethod()" will
never be affected by connect-non-signal because it always has the Q_SIGNALS
keyword.

BUG: 375239