Add and start using CLANG_CHECK_FUNC_NAME
Clazy checks so far have used the expensive getQualifiedNameAsString()
function which internally needs to build up a std::string from various
StringRefs, to check for fully-qualified function names.

The Clang-Tidy developers, however, recommend to first check the
function name via getName() (which returns a StringRef) and only upon
a successful match check the FQN, too.

Add a macro that implements this strategy. Can't use a real function,
since we needs to concatenate C string literals at compile time.

Not all instances of the pattern have been fixed. In a number of
places, the FQN is used to search in a list of names. These first need
to be decomposed.