Introduce autoconnect-slot, for detection of uic autoconnected slots
Warns for use of slots with the naming convention "on_btnRefresh_clicked",
which calls to the uic generated setupUi() method will automatically
connect to the corresponding widget's signals.

These connections are very fragile, and do not trigger compile time
or run time warnings when changes to the ui file (such as renaming,
removal of widgets, or changing widget type) cause the autoconnection
to fail.

Better to use explicit connections instead.