old-style-connect: Fix fixit bug when using Q_PRIVATE_SLOT
If the QObject has a signal with the same name of a Q_PRIVATE_SLOT
then the fixit would connect to the signal instead of the private
slot.

The fix is generic, if the user used the SLOT macro we simply
never rewrite to use a signal.

CCMAIL: dfaure@kdab.com