performance: Checks can now claim they can ignore includes
Most checks don't need to inspect the AST of their includes to
emit warnings in the .cpp file they're processing.

Some checks do, so this has to be done explicitly, case by case.
For now just added it to the checks pointed out by the profiler.

Useful for IDEs which are only showing inline warnings for the
current file.