Replace the REGISTER_* macros with auto-generated code
All checks are now registered in Checks.h, this file is auto generated
from the contents in checks.json.

This means checks.json is now the authority and this measure ensures
it's contents is correct, otherwise the unit-tests fail.

List of checks from Readme and anchor header will also be generated
from checks.json in the future.