Introduce cmake option: LINK_CLAZY_TO_LLVM
The default is ON, so nothing will change by default.

But to link against a static LLVM you probably want to set this
to OFF, specially if you're seeing runtime errors saying
"option foo registered more than once"

Will be using this option to make static builds work in the AppImage.