Make the clazy plugin also export fixes, not only clazy-standalone
Now fixits are always enabled (printed on screen, like clang and clang-tidy).
What the user can change is if they are exported to yaml files or not.

This makes the code much simpler and more similar to other tools.
The env variables to choose fixits are gone, you instead choose checks,
and pass --export-fixes= instead. (Or -Xclang -plugin-arg-clazy -Xclang export-fixes
for the plugin)