Fix clazy-standalone with llvm-9
This is a temporary fix so it doesn't block distros from using clazy
with llvm-9. And I'm not even sure others can reproduce, as it
might be a local compiler bug here. Debugged it quite a bit and
couldn't reach any conclusion.

Nikolai, it would be great if you could run:
	 $ cd tests/
	 $ clazy-standalone auto-unexpected-qstringbuilder/main.cpp

after having uncommented:
// llvm::errs() << optionsParser.getSourcePathList().size() << "\n";
in ClazyStandaloneMain.cpp

should not print 0

CCMAIL: nikolai.kosjar@qt.io