fully-qualified-moc-types: Explain why Q_PROPERTY warns on moc output only
Q_PROPERTY is a macro that doesn't expand to anything. It won't appear in
the AST. The properties only appear in the moc output.

It will warn that your Q_PROPERTY needs to be fully qualified when you do
a full build with clazy as the compiler, but probably not in QtCreator's
inline warnings.

Working at the Lexer level without a code model would not
only be hardwork but would be flaky with many corner cases and false-positives.

BUG: 423780