Fix Bug 442729 related to qt6-qlatin1stringchar-to-u fixit
QLatin1String should be replaced with u""_qs instead of u"".
QLatin1String cannot take a u""_qs as argument, or be constructed from it.
Also QLatin1Char cannot be constructed from a u''.
Consequently a fix can only be provided if a QString or QChar trace
has been found.

When a QLatin1Char/String call is found, the type of the parents
or checked in search for a QString or QChar type. In case of looking
for left over QLatin1Char/String calls nested in a
QLatin1Char/String call whose fix is not supported,
the type is looked for on the fly while scanning the
children within the outer QLatin1Char/String call.