tests: add -j/--jobs option for run_tests.py
This way it is possible to customize the number of parallel jobs to
spawn for the tests, instead of unconditionally using all available
CPUs. This can be helpful to run the tests without overloading the
system.

The default is still the number of available CPUs, so there is no
behaviour change by default.