Add sanitize-inline-keyword check
Emits a warning if the "inline" keyword is set on the definition
but not the declaration of a member method in an exported class.

This check sets Option_CanIgnoreIncludes; it's useful when using it with
clazy-standalone to avoid having duplicate fixes (which may end up
adding the inline keyword more than once to the same method).