Optimize checks for filesystem in tests
When we did determine this using cmake, we can reuse the value.
On my PC, this saves half a second for each test. With tests being run on CI, this adds up!

Even more, we checked the compilation in each test, even if the test did
not require the std::filesystem API