function-args-by-ref: Prefer emitting warning for non-trivial copying rather than size
In Qt5, QString has a size of 8, in Qt6, it is 24 due to the following
code:
```
class QString
    DataPointer d;
class QArrayDataPointer
    Data *d;
    T *ptr;
    qsizetype size;
```

Whereas in Qt5, we had `Data *d;` defined for the QString class, meaning
we had a pointer of a class that itself has only 3 pointers.

As a developer, the warning for non-trivially copyable classes is IMHO
more useful than the specific size in bytes. Especially since the action
one should take is exactly the same.