Remove Qt4 compatibility mode and last remaining check
By now, almost everyone should use Qt5 or Qt6. And even if there are old
projects, this one check will not be a must-have for them.

The Qt4 mode is also AFAIK untested and not actively maintained. Thus we
should focus on the relevant parts that have a consumer impact.