Make tests work when clang library is different from system default
If we have found the clang folder, we can resolve the executable from there.
In my case, I was compiling against clang16, but my system default was clang 14.
The stacktrace is really confusing and hard to understand. Using the
absolute path makes the log also better to understand and to reproduce
the test results.