Clang Tidy Plugin with Clazy Checks
This MR adds support for running clazy checks as part of clang-tidy. All checks must be enabled manually, since we can only emit warnings for checks that are explicitly or via wildcard enabled.

The architectural changes include here the lazy initialization of the clazy context, because clang-tidy has stricter lifecycle hooks than a normal clang::CompilerInstance forces us to. The warnings also need to be delegated to exactly the diagnostics engine of clang-tidy. Unlike in the clang plugin/standalone cases, the warnings are sorted before printing them out. This can lead to slightly different test results, adding the need for the ".expected.tidy" files.

Building the plugin can be controlled using the `CLAZY_BUILD_CLANG_TIDY` CMake option and the `run_test.py` script consumes the `--no-clang-tidy` flag. When using `ctest`, this is done automatically depending on the build option.

To keep logic in sync between the normal clazy and clang-tidy integration, a shared ClazyVisitHelper file was created. This takes care of calling the respective methods in `CheckBase` and updating the `ClazyContext` members.